JackieSharkSkin/foulingImages/pyproject.toml
[project]
name = "foulingimages"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "codechembook>=1.3.0",
    "lmfit>=1.3.4",
    "marimo>=0.19.9",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "scipy>=1.17.0",
    "statsmodels>=0.14.6",
]
